+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Always set status for new todo
 
+### Removed
+
+- Removed intuitive priority alias
+
 ## [v0.2.0] - 2025-07-25
 
 ### Added